requirements.txt
# ==============================================================================
# V100-Specific PyTorch Environment Configuration
# Target: CUDA 12.1 (cu121) — Compatible with NVIDIA Volta (sm_70, compute 7.0)
# Requirement: NVIDIA Driver >= 525.60
#
# NOTE:
# CUDA 12.4+ builds may introduce sm_70 kernel incompatibilities on V100 GPUs.
# Avoid cu124+ unless driver compatibility is explicitly verified.
# ==============================================================================

--index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://pypi.org/simple


# ------------------------------------------------------------------------------
# Core PyTorch Framework (CUDA 12.1 Builds)
# ------------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0


# ------------------------------------------------------------------------------
# CUDA 12.1 Runtime Libraries
# ------------------------------------------------------------------------------
nvidia-cuda-runtime-cu12>=12.1.105
nvidia-cuda-nvrtc-cu12>=12.1.105        # Required for JIT kernel compilation (Triton dependency)
nvidia-cuda-cupti-cu12>=12.1.105        # Profiling and performance counters
nvidia-cudnn-cu12>=8.9.2.26             # cuDNN 8.x (fully supported on sm_70)
nvidia-cublas-cu12>=12.1.3.1            # BLAS backend for linear algebra
nvidia-cufft-cu12>=11.0.2.54            # FFT acceleration
nvidia-curand-cu12>=10.3.2.106          # GPU random number generation
nvidia-cusolver-cu12>=11.4.5.107        # Dense and sparse linear solvers
nvidia-cusparse-cu12>=12.1.0.106        # Sparse matrix operations
nvidia-nccl-cu12>=2.18.1                # Multi-GPU collectives (NVLink on V100 SXM2)
nvidia-nvtx-cu12>=12.1.105              # NVTX profiling markers


# ------------------------------------------------------------------------------
# Quantum / Photonic Simulation
# ------------------------------------------------------------------------------
perceval-quandela>=0.11.0


# ------------------------------------------------------------------------------
# Machine Learning / Data Science Stack
# ------------------------------------------------------------------------------
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.64.0
scikit-learn>=1.2.0
datasets
merlinquantum